Fix issues in building

+++ projects/btldr/libs/shmem/api.h
@@ -6,7 +6,7 @@
  * Bootflags
  */
 #define UPDATE_REQUESTED (uint32_t)(1 << 0x000U)
-#define IMAGE_IS_VALID   (uint32_t)(1 << 0x001U)
+#define IMAGE_IS_VALID (uint32_t)(1 << 0x001U)
 
 union updater_version {
     const uint8_t full;
@@ -29,7 +29,7 @@ void bootflag_clear(uint8_t flag);
 /*
  * Get the value of the specified bootflag
  */
-bool bootflag_get(uint8_t flag);
+bool bootflag_get(uint32_t flag);
 
 /*
  * Used to get the 1-byte updater version from EEPROM
